chore: separate file for package resolution algorithm
